Add more Unit-tests

+++ internal/repository/postgres/postgres_todo_item.go
@@ -92,8 +92,9 @@ func (r *TodoItem) Update(userId, itemId int, input entity.UpdateItemInput) erro
 
 	setQuery := strings.Join(setValues, ", ")
 
-	query := fmt.Sprintf("UPDATE %s AS ti SET %s FROM %s AS ul, %s AS li "+
-		"WHERE ti.id = li.item_id AND li.list_id = ul.list_id AND ul.user_id = $%d AND ti.id = $%d",
+	query := fmt.Sprintf(`UPDATE %s AS ti SET %s 
+									FROM %s AS ul, %s AS li 
+									WHERE ti.id = li.item_id AND li.list_id = ul.list_id AND ul.user_id = $%d AND ti.id = $%d`,
 		todoItemsTable, setQuery, usersListsTable, listsItemsTable, argId, argId+1)
 
 	args = append(args, userId, itemId)
@@ -104,8 +105,7 @@ func (r *TodoItem) Update(userId, itemId int, input entity.UpdateItemInput) erro
 }
 
 func (r *TodoItem) Delete(userId, itemId int) error {
-	query := fmt.Sprintf("DELETE FROM %s AS ti USING %s as ul, %s as li "+
-		"WHERE  ti.id = li.item_id AND li.list_id = ul.list_id AND ul.user_id = $1 AND ti.id = $2;",
+	query := fmt.Sprintf(`DELETE FROM %s AS ti USING %s as ul, %s as li WHERE  ti.id = li.item_id AND li.list_id = ul.list_id AND ul.user_id = $1 AND ti.id = $2;`,
 		todoItemsTable, usersListsTable, listsItemsTable)
 	_, err := r.db.Exec(query, userId, itemId)
